--- libs/application/templates/announcement-of-death/src/fields/Done/AnnouncementCompleteImage.tsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-import AOD from '../../assets/AOD'
-import { Box } from '@island.is/island-ui/core'
-
-export const AnnouncementCompleteImage = () => {
-  return (
-    <Box marginTop={3} marginBottom={10} display="flex" justifyContent="center">
-      <AOD />
-    </Box>
-  )
-}
