+++ libs/application/templates/announcement-of-death/src/fields/Done/index.tsx
@@ -1,3 +1,12 @@
-export { Completed } from './Completed'
-export { Delegated } from './Delegated'
-export { AnnouncementCompleteImage } from './AnnouncementCompleteImage'
+import React from 'react'
+
+import AOD from '../../assets/AOD'
+import { Box } from '@island.is/island-ui/core'
+
+export const Done = () => {
+  return (
+    <Box marginTop={6} marginBottom={10} display="flex" justifyContent="center">
+      <AOD />
+    </Box>
+  )
+}
